Start work to support Encodable parameters for requests
