+++ Sources/SwiftTwitch/Classes/Helping Extensions/DictionaryExtensions.swift
@@ -22,3 +22,16 @@ extension Dictionary {
         self[key] = nonNilValue
     }
 }
+
+extension Dictionary where Key == String, Value == Any {
+
+    /// Converts the `Dictionary` to its `Data` representation.
+    ///
+    /// - Returns: The `Data` representation of the `Dictionary`.
+    internal func getAsData() -> Data? {
+        guard let jsonData = try? JSONSerialization.data(withJSONObject: self) else {
+            return nil
+        }
+        return jsonData
+    }
+}
