+++ Sources/SwiftTwitch/Classes/Helping Extensions/EncodableExtensions.swift
@@ -1,0 +1,15 @@
+//
+//  EncodableExtensions.swift
+//  
+//
+//  Created by Kevin Romero Peces-Barba on 14/07/2020.
+//
+
+import Foundation
+
+extension Encodable {
+  var dictionary: [String: Any]? {
+    guard let data = try? JSONEncoder().encode(self) else { return nil }
+    return (try? JSONSerialization.jsonObject(with: data, options: .allowFragments)).flatMap { $0 as? [String: Any] }
+  }
+}
