+++ Sources/SwiftTwitch/Classes/Requests/EmptyParams.swift
@@ -1,0 +1,10 @@
+//
+//  File.swift
+//  
+//
+//  Created by Kevin Romero Peces-Barba on 14/07/2020.
+//
+
+import Foundation
+
+struct EmptyParams: Codable {}
