+++ Sources/SwiftTwitch/Classes/Requests/Users/UpdateUserParams.swift
@@ -1,0 +1,15 @@
+//
+//  UpdateUserParams.swift
+//  
+//
+//  Created by Kevin Romero Peces-Barba on 14/07/2020.
+//
+
+import Foundation
+
+public struct UpdateUserParams: Codable {
+
+    /// `description`: The description to set for the user.
+    let description: String?
+
+}
