+++ Sources/SwiftTwitch/Classes/Requests/WebRequestKeys.swift
@@ -1,0 +1,99 @@
+//
+//  WebRequestKeys.swift
+//  
+//
+//  Created by Kevin Romero Peces-Barba on 14/07/2020.
+//
+
+import Foundation
+
+/// `WebRequestKeys` define the web request keys for both resolving results and sending requests
+/// for the New Twitch API.
+internal struct WebRequestKeys {
+    static let ability = "ability"
+    static let after = "after"
+    static let before = "before"
+    static let boxArtURL = "box_art_url"
+    static let broadcaster = "broadcaster"
+    static let broadcasterId = "broadcaster_id"
+    static let broadcasterName = "broadcaster_name"
+    static let broadcasterType = "broadcaster_type"
+    static let canActivate = "can_activate"
+    static let classKey = "class"
+    static let communityId = "community_id"
+    static let communityIds = "community_ids"
+    static let count = "count"
+    static let createdAt = "created_at"
+    static let creatorId = "creator_id"
+    static let creatorName = "creator_name"
+    static let cursor = "cursor"
+    static let data = "data"
+    static let dateRange = "date_range"
+    static let description = "description"
+    static let displayName = "display_name"
+    static let duration = "duration"
+    static let editURL = "edit_url"
+    static let email = "email"
+    static let embedURL = "embed_url"
+    static let endedAt = "ended_at"
+    static let extensionId = "extension_id"
+    static let first = "first"
+    static let followedAt = "followed_at"
+    static let fromId = "from_id"
+    static let fromName = "from_name"
+    static let gameId = "game_id"
+    static let hasDelay = "has_delay"
+    static let hearthstone = "hearthstone"
+    static let hero = "hero"
+    static let id = "id"
+    static let language = "language"
+    static let login = "login"
+    static let manifestId = "manifest_id"
+    static let markers = "markers"
+    static let name = "name"
+    static let offlineImageURL = "offline_image_url"
+    static let opponent = "opponent"
+    static let overwatch = "overwatch"
+    static let pagination = "pagination"
+    static let period = "period"
+    static let positionSeconds = "position_seconds"
+    static let profileImageURL = "profile_image_url"
+    static let publishedAt = "published_at"
+    static let rank = "rank"
+    static let role = "role"
+    static let score = "score"
+    static let startedAt = "started_at"
+    static let sort = "sort"
+    static let thumbnailURL = "thumbnail_url"
+    static let title = "title"
+    static let toId = "to_id"
+    static let toName = "to_name"
+    static let total = "total"
+    static let type = "type"
+    static let url = "url"
+    static let userId = "user_id"
+    static let userLogin = "user_login"
+    static let userName = "user_name"
+    static let version = "version"
+    static let videoId = "video_id"
+    static let videos = "videos"
+    static let viewable = "viewable"
+    static let viewCount = "view_count"
+    static let viewerCount = "viewer_count"
+}
+
+internal struct WebRequestKeysV5 {
+    static let streamType = "stream_type"
+    static let limit = "limit"
+    static let offset = "offset"
+    static let streams = "streams"
+    static let total = "_total"
+    static let id = "_id"
+    static let channelID = "channel._id"
+    static let channelName = "channel.name"
+    static let viewerCount = "viewers"
+    static let title = "channel.status"
+    static let createdAt = "created_at"
+    static let language = "channel.language"
+    static let thumbnailURL = "preview.template"
+}
